# Improve the JSON parser performance.