--- a/pkg/logql/log/parser_hints.go
+++ b/pkg/logql/log/parser_hints.go
@@ -25,16 +25,11 @@ type ParserHint interface {
 	//		 sum(rate({app="foo"} | json [5m]))
 	// We don't need to extract any labels from the log line.
 	NoLabels() bool
-	Keys() []string
 }
 
 type parserHint struct {
 	noLabels       bool
 	requiredLabels []string
-}
-
-func (p *parserHint) Keys() []string {
-	return p.requiredLabels
 }
 
 func (p *parserHint) ShouldExtract(key string) bool {