Sorry I couldn't help myself and worked a bit on the JSON parser by using another library that allows way less allocations.

On top of this, the parser code is also much simpler to understand.

I realize that we could possibly also intern/cache (limited obviously) extracted JSON value per key. I would do it per key because otherwise, a high cardinality key can take over the intern/cache which will make it worthless. Happy to walk someone through the idea, I think this would be a big jump too.

Would be great to run and test this in our dev environment using some JSON queries before releasing it.

```
❯ benchstat before.txt after.txt                                                                                                              
name                             old time/op    new time/op    delta
_Parser/json/no_labels_hints-10    2.53µs ± 0%    2.16µs ± 2%  -14.59%  (p=0.008 n=5+5)
_Parser/json/labels_hints-10       1.93µs ± 0%    1.47µs ± 0%  -23.96%  (p=0.008 n=5+5)

name                             old alloc/op   new alloc/op   delta
_Parser/json/no_labels_hints-10      656B ± 0%      280B ± 0%  -57.32%  (p=0.008 n=5+5)
_Parser/json/labels_hints-10         512B ± 0%      176B ± 0%  -65.62%  (p=0.008 n=5+5)

name                             old allocs/op  new allocs/op  delta
_Parser/json/no_labels_hints-10      46.0 ± 0%      18.0 ± 0%  -60.87%  (p=0.008 n=5+5)
_Parser/json/labels_hints-10         39.0 ± 0%      12.0 ± 0%  -69.23%  (p=0.008 n=5+5)
```