@vlad-diachenko 2023-01-13
@cyriltovena should we mention such changes in the changelog? I believe that it might worth to mention because it improves json parser performance a lot ... ))