@cyriltovena 2023-01-13
> @cyriltovena should we mention such changes in the changelog? I believe that it might worth to mention because it improves json parser performance a lot ... ))

Sure however it would be better to try it in a cluster see the real difference. May be you can compare before and after ? I'd love to know.